+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/Users/Munchic/.local/share/virtualenvs/unsam-I7mHUPgb/bin/python"
+}
